Separate addon scanning from agent helper concerns
The source scanner has an independent AST/XML boundary and should remain
usable without coupling its implementation to the broader agent helper
module. Compatibility re-exports preserve existing callers while the new
module boundary is covered directly.

Constraint: Preserve both addon-scanner and field-ranking compatibility re-exports after rebasing.
Confidence: high
Scope-risk: narrow
Directive: Keep addon scanner helpers isolated from MCP surface modules and retain agent_tools compatibility imports.
Tested: 910 pytest tests; Ruff; mypy; Black; import-linter; git diff checks.
Not-tested: Live Odoo addon directory smoke test.
